+++ packages/patternfly-4/react-core/src/components/Card/Card.tsx
@@ -11,6 +11,8 @@ export interface CardProps extends React.HTMLProps<HTMLDivElement> {
   component?: React.ReactNode;
   /** Modifies the card to include hover styles on :hover */
   isHoverable?: boolean;
+  /** Modifies the card to include compact styling */
+  isCompact?: boolean; 
 }
 
 export const Card: React.FunctionComponent<CardProps> = ({
@@ -18,11 +20,13 @@ export const Card: React.FunctionComponent<CardProps> = ({
   className = '',
   component = 'article',
   isHoverable = false, 
+  isCompact = false,
   ...props
 }: CardProps) => {
   const Component = component as any;
   return (
-  <Component className={css(styles.card, isHoverable && styles.modifiers.hoverable, className)} {...props}>
+  <Component className={css(styles.card, isHoverable && styles.modifiers.hoverable, isCompact && styles.modifiers.compact, className)} {...props}>
     {children}
   </Component>
 )};
+