+++ packages/patternfly-4/react-core/src/components/Progress/index.d.ts
@@ -1,1 +1,0 @@
-export { default as Progress, ProgressVariant, ProgressMeasureLocation, ProgressSize, ProgressProps } from './Progress';