+++ packages/patternfly-4/react-core/src/components/Progress/index.ts
@@ -1,0 +1,2 @@
+export * from './Progress';
+export { ProgressMeasureLocation, ProgressVariant } from './ProgressContainer';