+++ packages/patternfly-4/react-core/src/components/Radio/Radio.tsx
@@ -3,7 +3,8 @@ import styles from '@patternfly/react-styles/css/components/Radio/radio';
 import { css, getModifier } from '@patternfly/react-styles';
 import { Omit } from '../../helpers/typeUtils';
 
-export interface RadioProps extends Omit<React.HTMLProps<HTMLInputElement>, 'disabled' | 'label' | 'onChange' | 'type'> {
+export interface RadioProps
+  extends Omit<React.HTMLProps<HTMLInputElement>, 'disabled' | 'label' | 'onChange' | 'type'> {
   /** Additional classes added to the radio. */
   className?: string;
   /** Id of the radio. */
@@ -28,8 +29,9 @@ export class Radio extends React.Component<RadioProps> {
   static defaultProps = {
     className: '',
     isDisabled: false,
-    isValid: true
-  }
+    isValid: true,
+    onChange: Function.prototype
+  };
 
   constructor(props: RadioProps) {
     super(props);
@@ -41,7 +43,7 @@ export class Radio extends React.Component<RadioProps> {
 
   handleChange = (event: React.FormEvent<HTMLInputElement>) => {
     this.props.onChange(event.currentTarget.checked, event);
-  }
+  };
 
   render() {
     const {
@@ -60,17 +62,20 @@ export class Radio extends React.Component<RadioProps> {
       <div className={css(styles.radio, className)}>
         <input
           {...props}
-          aria-label={label ? undefined : ariaLabel}
           className={css(styles.radioInput)}
           type="radio"
           onChange={this.handleChange}
           aria-invalid={!isValid}
           disabled={isDisabled}
           checked={checked || isChecked}
-          defaultChecked={isChecked ? undefined : defaultChecked}
+          {...!isChecked && { defaultChecked }}
+          {...!label && { 'aria-label': ariaLabel }}
         />
         {label && (
-          <label className={css(styles.radioLabel, getModifier(styles, isDisabled && 'disabled' as any))} htmlFor={props.id}>
+          <label
+            className={css(styles.radioLabel, getModifier(styles, isDisabled && ('disabled' as any)))}
+            htmlFor={props.id}
+          >
             {label}
           </label>
         )}