+++ packages/patternfly-4/react-core/src/components/Select/index.js
@@ -1,5 +1,0 @@
-export { default as Select } from './Select';
-export { default as SelectOption } from './SelectOption';
-export { SelectContext, SelectVariant } from './selectConstants';
-export { default as CheckboxSelectGroup } from './CheckboxSelectGroup';
-export { default as CheckboxSelectOption } from './CheckboxSelectOption';