+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/CardDemo/CardDemo.tsx
@@ -4,11 +4,19 @@ import { Card, CardHeader, CardBody, CardFooter } from '@patternfly/react-core';
 export class CardDemo extends React.Component {
   render() {
     return (
-      <Card isHoverable={true}>
-        <CardHeader>Header</CardHeader>
-        <CardBody>Body</CardBody>
-        <CardFooter>Footer</CardFooter>
-      </Card>
+      <React.Fragment> 
+        <Card isHoverable>
+          <CardHeader>Header</CardHeader>
+          <CardBody>Body</CardBody>
+          <CardFooter>Footer</CardFooter>
+        </Card>
+        <br></br> 
+        <Card isCompact>
+          <CardHeader>Header</CardHeader>
+          <CardBody>Body</CardBody>
+          <CardFooter>Footer</CardFooter>
+        </Card>
+      </React.Fragment>
     )
   }
 }