+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/ProgressDemo/ProgressDemo.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Progress } from '@patternfly/react-core';
+
+export class ProgressDemo extends React.Component {
+  render() {
+    return <Progress id="simple-progress" value={33} title="Simple progress" />;
+  }
+}