+++ packages/patternfly-4/react-topology/CHANGELOG.md
@@ -3,6 +3,89 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.18...@patternfly/react-topology@2.4.19) (2019-07-01)
+
+
+### Bug Fixes
+
+* **tooltip:** revert tippy2 upgrade ([#2326](https://github.com/patternfly/patternfly-react/issues/2326)) ([#2414](https://github.com/patternfly/patternfly-react/issues/2414)) ([3e919e3](https://github.com/patternfly/patternfly-react/commit/3e919e3))
+
+
+
+
+
+## [2.4.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.17...@patternfly/react-topology@2.4.18) (2019-07-01)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.16...@patternfly/react-topology@2.4.17) (2019-07-01)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.15...@patternfly/react-topology@2.4.16) (2019-07-01)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.14...@patternfly/react-topology@2.4.15) (2019-07-01)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.13...@patternfly/react-topology@2.4.14) (2019-07-01)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.12...@patternfly/react-topology@2.4.13) (2019-06-28)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.11...@patternfly/react-topology@2.4.12) (2019-06-27)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.10...@patternfly/react-topology@2.4.11) (2019-06-27)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
+## [2.4.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.9...@patternfly/react-topology@2.4.10) (2019-06-27)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
 ## [2.4.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.4.8...@patternfly/react-topology@2.4.9) (2019-06-27)
 
 **Note:** Version bump only for package @patternfly/react-topology